+++ src/sdk.rs
@@ -1,6 +1,6 @@
 use crate::{
 	error::ClientError, rpc, transactions::Transactions, ABlock, ABlocksClient, AOnlineClient, AStorageClient,
-	AvailHeader,
+	AvailHeader, TransactionState,
 };
 use primitive_types::H256;
 use std::{fmt::Debug, time::Duration};
@@ -84,6 +84,8 @@ impl Debug for SDK {
 
 pub async fn reconnecting_api(endpoint: &str) -> Result<Client, ClientError> {
 	let rpc_client = ReconnectingRpcClient::builder()
+		.max_request_size(512 * 1024 * 1024)
+		.max_response_size(512 * 1024 * 1024)
 		.retry_policy(
 			ExponentialBackoff::from_millis(1000)
 				.max_delay(Duration::from_secs(3))
@@ -164,12 +166,21 @@ impl Client {
 		rpc::chain::get_header(self, Some(at)).await
 	}
 
+	pub async fn block_hash(&self, block_height: u32) -> Result<H256, subxt::Error> {
+		rpc::chain::get_block_hash(self, Some(block_height)).await
+	}
+
 	pub async fn best_block_hash(&self) -> Result<H256, subxt::Error> {
 		rpc::chain::get_block_hash(self, None).await
 	}
 
 	pub async fn finalized_block_hash(&self) -> Result<H256, subxt::Error> {
 		rpc::chain::get_finalized_head(self).await
+	}
+
+	pub async fn block_number(&self, block_hash: H256) -> Result<u32, subxt::Error> {
+		let header = rpc::chain::get_header(self, Some(block_hash)).await?;
+		Ok(header.number)
 	}
 
 	pub async fn best_block_number(&self) -> Result<u32, subxt::Error> {
@@ -181,6 +192,14 @@ impl Client {
 		let block_hash = self.finalized_block_hash().await?;
 		let header = rpc::chain::get_header(self, Some(block_hash)).await?;
 		Ok(header.number)
+	}
+
+	pub async fn transaction_state(
+		&self,
+		tx_hash: &H256,
+		finalized: bool,
+	) -> Result<Vec<TransactionState>, subxt::Error> {
+		rpc::transaction::state(self, tx_hash, finalized).await
 	}
 }